+++ libs/sdk-core/src/lib/db/kb/kb.models.ts
@@ -1,6 +1,5 @@
 import type { Observable } from 'rxjs';
 import type { IResource, LinkField, Resource, UserMetadata } from '../resource';
-import { FieldId, IFieldData } from '../resource';
 import type { FileMetadata, FileWithMetadata, UploadResponse, UploadStatus } from '../upload';
 import type { Search, SearchOptions } from '../search';
 
@@ -72,21 +71,8 @@ export interface IKnowledgeBox extends IKnowledgeBoxCreation {
   getLabels(): Observable<LabelSets>;
 
   getResource(uuid: string, show?: ResourceProperties[], extracted?: ExtractedDataTypes[]): Observable<IResource>;
+
   getResourceBySlug(slug: string, show?: ResourceProperties[], extracted?: ExtractedDataTypes[]): Observable<IResource>;
-
-  getResourceField(
-    rid: string,
-    fieldId: FieldId,
-    show?: ResourceFieldProperties[],
-    extracted?: ExtractedDataTypes[],
-  ): Observable<IFieldData & FieldId>;
-
-  getResourceFieldBySlug(
-    slug: string,
-    fieldId: FieldId,
-    show?: ResourceFieldProperties[],
-    extracted?: ExtractedDataTypes[],
-  ): Observable<IFieldData & FieldId>;
 
   search(query: string, features?: Search.Features[], options?: SearchOptions): Observable<Search.Results>;